+++ resources/views/pages/admin/knowledge/reviews.blade.php
@@ -1,0 +1,5 @@
+<x-layouts::app :title="__('Validation Review Queue')">
+    <section class="w-full">
+        <livewire:admin.knowledge.review-queue />
+    </section>
+</x-layouts::app>